Extract the Auth and Token logic to an auth module
